Split persistence_spec.rb into multiple files in persistence/ directory
